refactor(likeButton): Managed files
