+++ likebutton/src/App.tsx
@@ -1,9 +1,7 @@
-import LikeButton from "./pages/likeButton"
+import LikeButton from "./components/likeButton"
 import './App.css'
 
 function App() {
-
-
   return (
     <>
       <LikeButton />
