+++ likebutton/src/components/likeButton.module.css
@@ -12,7 +12,7 @@
     color: red;
 }
 
-.btn-liked{
+.btnLiked{
     border-radius: 29px;
     font-size: 25px;
     border: 3px solid red;
