+++ likebutton/src/components/likeButton.tsx
@@ -1,6 +1,6 @@
-import "../css/likeButton.css";
 import { useState } from "react";
 import "bootstrap-icons/font/bootstrap-icons.css";
+import styles from "./likeButton.module.css";
 const LikeButton = () => {
   const [liked, setLiked] = useState(false);
 
@@ -10,7 +10,7 @@ const LikeButton = () => {
   return (
     <>
       <div>
-        <button className={liked ? "btn-liked" : "btn"} onClick={onLiked}>
+        <button className={liked ? styles.btnLiked : styles.btn} onClick={onLiked}>
           <i className="bi bi-heart-fill" style={{color: liked ? "white" : "red"}}></i> Like
         </button>
       </div>
